+++ src/main/groovy/com/getkeepsafe/dexcount/DexMethodCountTask.groovy
@@ -17,17 +17,14 @@
 package com.getkeepsafe.dexcount
 
 import com.android.annotations.Nullable
+import com.android.annotations.VisibleForTesting
 import com.android.build.gradle.api.BaseVariantOutput
-import com.android.dexdeps.HasDeclaringClass
-import com.android.dexdeps.Output
 import groovy.transform.stc.ClosureParams
 import groovy.transform.stc.SimpleType
 import org.gradle.api.DefaultTask
 import org.gradle.api.logging.LogLevel
-import org.gradle.api.tasks.OutputFile
+import org.gradle.api.logging.Logger
 import org.gradle.api.tasks.TaskAction
-import org.gradle.logging.StyledTextOutput
-import org.gradle.logging.StyledTextOutputFactory
 
 class DexMethodCountTask extends DefaultTask {
     /**
@@ -43,11 +40,9 @@ class DexMethodCountTask extends DefaultTask {
     @Nullable
     def File mappingFile
 
-    @OutputFile
     def File outputFile
-
-    @OutputFile
     def File summaryFile
+    def File chartDir
 
     def DexMethodCountExtension config
 
@@ -56,17 +51,39 @@ class DexMethodCountTask extends DefaultTask {
     def long treegenTime
     def long outputTime
 
+    def boolean isInstantRun
+
     @TaskAction
     void countMethods() {
-        generatePackageTree()
-        printSummary()
-        printFullTree()
-        printTaskDiagnosticData()
+        try {
+            printPreamble()
+            generatePackageTree()
+            printSummary()
+            printFullTree()
+            printChart()
+            printTaskDiagnosticData()
+        } catch (DexCountException e) {
+            withStyledOutput() { out ->
+                out.error("Error counting dex methods. Please contact the developer at https://github.com/KeepSafe/dexcount-gradle-plugin/issues", e)
+            }
+        }
     }
 
     static def percentUsed(int count) {
         def used = ((double) count / MAX_DEX_REFS) * 100.0
         return sprintf("%.2f", used)
+    }
+
+    def printPreamble() {
+        if (config.printVersion) {
+            def projectName = getClass().package.implementationTitle
+            def projectVersion = getClass().package.implementationVersion
+
+            withStyledOutput() { out ->
+                out.lifecycle("Dexcount name:    $projectName")
+                out.lifecycle("Dexcount version: $projectVersion")
+            }
+        }
     }
 
     /**
@@ -75,14 +92,24 @@ class DexMethodCountTask extends DefaultTask {
      */
     def printSummary() {
         def filename = apkOrDex.outputFile.name
-        withStyledOutput(StyledTextOutput.Style.Info) { out ->
+
+        if (isInstantRun) {
+            withStyledOutput() { out ->
+                out.warn("Warning: Instant Run build detected!  Instant Run does not run Proguard; method counts may be inaccurate.")
+            }
+        }
+
+        withStyledOutput() { out ->
             def percentMethodsUsed = percentUsed(tree.methodCount)
             def percentFieldsUsed = percentUsed(tree.fieldCount)
 
-            out.println("Total methods in ${filename}: ${tree.methodCount} ($percentMethodsUsed% used)")
-            out.println("Total fields in ${filename}:  ${tree.fieldCount} ($percentFieldsUsed% used)")
-            out.println("Methods remaining in ${filename}: ${MAX_DEX_REFS - tree.methodCount}")
-            out.println("Fields remaining in ${filename}:  ${MAX_DEX_REFS - tree.fieldCount}")
+            def methodsRemaining = Math.max(MAX_DEX_REFS - tree.methodCount, 0)
+            def fieldsRemaining = Math.max(MAX_DEX_REFS - tree.fieldCount, 0)
+
+            out.warn("Total methods in ${filename}: ${tree.methodCount} ($percentMethodsUsed% used)")
+            out.warn("Total fields in ${filename}:  ${tree.fieldCount} ($percentFieldsUsed% used)")
+            out.warn("Methods remaining in ${filename}: $methodsRemaining")
+            out.warn("Fields remaining in ${filename}:  $fieldsRemaining")
         }
 
         if (summaryFile != null) {
@@ -98,24 +125,55 @@ class DexMethodCountTask extends DefaultTask {
                 appendableStream.println(counts);
             }
         }
+
+        if (getPrintOptions().teamCityIntegration || config.teamCitySlug != null) {
+            withStyledOutput() { out ->
+                def prefix = "Dexcount"
+                if (config.teamCitySlug != null) {
+                    def slug = config.teamCitySlug.replace(' ', '_') // Not sure how TeamCity would handle spaces?
+                    prefix = "${prefix}_${slug}"
+                }
+
+                printTeamCityStatisticValue(out, "${prefix}_${apkOrDex.name}_MethodCount", tree.methodCount)
+                printTeamCityStatisticValue(out, "${prefix}_${apkOrDex.name}_FieldCount", tree.fieldCount)
+            }
+        }
+    }
+
+    /**
+     * Reports to Team City statistic value
+     * Doc: https://confluence.jetbrains.com/display/TCD9/Build+Script+Interaction+with+TeamCity#BuildScriptInteractionwithTeamCity-ReportingBuildStatistics
+     */
+    def printTeamCityStatisticValue(Logger out, String key, int value) {
+        out.lifecycle("##teamcity[buildStatisticValue key='${key}' value='${value}']")
     }
 
     /**
      * Prints the package tree to the usual outputs/dexcount/variant.txt file.
      */
     def printFullTree() {
-        if (outputFile != null) {
-            outputFile.parentFile.mkdirs()
-            outputFile.createNewFile()
-            outputFile.withOutputStream { stream ->
-                def appendableStream = new PrintStream(stream)
-                print(tree, appendableStream)
-                appendableStream.flush()
-                appendableStream.close()
-            }
+        printToFile(outputFile) { PrintStream out ->
+            print(tree, out)
+        }
+        outputTime = System.currentTimeMillis()
+    }
+
+    /**
+     * Prints the package tree as chart to the outputs/dexcount/${variant}Chart directory.
+     */
+    def printChart() {
+        def printOptions = getPrintOptions()
+        printOptions.includeClasses = true
+        printToFile(new File(chartDir, "data.js")) { PrintStream out ->
+            out.print("var data = ")
+            tree.printJson(out, printOptions);
         }
 
-        outputTime = System.currentTimeMillis()
+        ["chart-builder.js", "d3.v3.min.js", "index.html", "styles.css"].each { String resourceName ->
+            def resource = getClass().getResourceAsStream("/com/getkeepsafe/dexcount/" + resourceName);
+            def targetFile = new File(chartDir, resourceName)
+            targetFile.write resource.text
+        }
     }
 
     /**
@@ -125,56 +183,69 @@ class DexMethodCountTask extends DefaultTask {
     def printTaskDiagnosticData() {
         // Log the entire package list/tree at LogLevel.DEBUG, unless
         // verbose is enabled (in which case use the default log level).
-        def level = config.verbose ? null : LogLevel.DEBUG
+        def level = config.verbose ? LogLevel.LIFECYCLE : LogLevel.DEBUG
 
-        withStyledOutput(StyledTextOutput.Style.Info, level) { out ->
-            print(tree, out)
+        withStyledOutput() { out ->
+            def strBuilder = new StringBuilder()
+            print(tree, strBuilder)
 
-            out.format("\n\nTask runtimes:\n")
-            out.format("--------------\n")
-            out.format("parsing:    ${ioTime - startTime} ms\n")
-            out.format("counting:   ${treegenTime - ioTime} ms\n")
-            out.format("printing:   ${outputTime - treegenTime} ms\n")
-            out.format("total:      ${outputTime - startTime} ms\n")
+            out.log(level, strBuilder.toString())
+            out.log(level, "\n\nTask runtimes:")
+            out.log(level, "--------------")
+            out.log(level, "parsing:    ${ioTime - startTime} ms")
+            out.log(level, "counting:   ${treegenTime - ioTime} ms")
+            out.log(level, "printing:   ${outputTime - treegenTime} ms")
+            out.log(level, "total:      ${outputTime - startTime} ms")
         }
     }
 
-    def print(PackageTree tree, Appendable writer) {
-        tree.print(writer, config.format, getPrintOptions())
+    def print(PackageTree tree, Appendable out) {
+        tree.print(out, config.format, getPrintOptions())
     }
 
-    private void withStyledOutput(
-            StyledTextOutput.Style style,
-            LogLevel level = null,
-            @ClosureParams(value = SimpleType, options = ['org.gradle.logging.StyledTextOutput']) Closure closure) {
-        def factory = services.get(StyledTextOutputFactory)
-        def output = level == null ? factory.create('dexcount') : factory.create('dexcount', level)
+    private void withStyledOutput(@ClosureParams(value = SimpleType, options = ['org.gradle.api.logging.Logger']) Closure closure) {
+        // TODO: Actually make this stylized when we have our own solution: https://github.com/KeepSafe/dexcount-gradle-plugin/issues/124
+        closure(getLogger())
+    }
 
-        closure(output.withStyle(style))
+    private void printToFile(
+            File file,
+            @ClosureParams(value = SimpleType, options = ['java.io.PrintStream']) Closure closure) {
+        if (outputFile != null) {
+            file.parentFile.mkdirs()
+            file.createNewFile()
+            file.withOutputStream { stream ->
+                def out = new PrintStream(stream)
+                closure(out)
+                out.flush()
+                out.close()
+            }
+        }
     }
 
     /**
      * Creates a new PackageTree and populates it with the method and field
      * counts of the current dex/apk file.
      */
-    private def generatePackageTree() {
+    @VisibleForTesting
+    def generatePackageTree() {
         startTime = System.currentTimeMillis()
 
         // Create a de-obfuscator based on the current Proguard mapping file.
         // If none is given, we'll get a default mapping.
         def deobs = getDeobfuscator()
 
-        def dataList = DexFile.extractDexData(apkOrDex.outputFile)
+        def dataList = DexFile.extractDexData(apkOrDex.outputFile, config.dxTimeoutSec)
 
         ioTime = System.currentTimeMillis()
         try {
-            tree = new PackageTree()
+            tree = new PackageTree(deobs)
 
-            refListToClassNames(dataList*.getMethodRefs(), deobs).each {
+            dataList*.getMethodRefs().flatten().each {
                 tree.addMethodRef(it)
             }
 
-            refListToClassNames(dataList*.getFieldRefs(), deobs).each {
+            dataList*.getFieldRefs().flatten().each {
                 tree.addFieldRef(it)
             }
         } finally {
@@ -182,22 +253,8 @@ class DexMethodCountTask extends DefaultTask {
         }
 
         treegenTime = System.currentTimeMillis()
-    }
 
-    static refListToClassNames(List<List<HasDeclaringClass>> refs, Deobfuscator deobfuscator) {
-        return refs.flatten().collect { ref ->
-            def descriptor = ref.getDeclClassName()
-            def dot = Output.descriptorToDot(descriptor)
-            dot = deobfuscator.deobfuscate(dot)
-            if (dot.indexOf('.') == -1) {
-                // Classes in the unnamed package (e.g. primitive arrays)
-                // will not appear in the output in the current PackageTree
-                // implementation if classes are not included.  To work around,
-                // we make an artificial package named "<unnamed>".
-                dot = "<unnamed>." + dot
-            }
-            return dot
-        }
+        isInstantRun = dataList.any { it.isInstantRun }
     }
 
     private def getPrintOptions() {
@@ -205,15 +262,17 @@ class DexMethodCountTask extends DefaultTask {
                 includeMethodCount: true,
                 includeFieldCount: config.includeFieldCount,
                 includeTotalMethodCount: config.includeTotalMethodCount,
+                teamCityIntegration: config.teamCityIntegration,
                 orderByMethodCount: config.orderByMethodCount,
                 includeClasses: config.includeClasses,
-                printHeader: true)
+                printHeader: true,
+                maxTreeDepth: config.maxTreeDepth)
     }
 
     private def getDeobfuscator() {
         if (mappingFile != null && !mappingFile.exists()) {
-            withStyledOutput(StyledTextOutput.Style.Normal, LogLevel.DEBUG) {
-                it.println("Mapping file specified at ${mappingFile.absolutePath} does not exist, assuming output is not obfuscated.")
+            withStyledOutput() {
+                it.debug("Mapping file specified at ${mappingFile.absolutePath} does not exist, assuming output is not obfuscated.")
             }
             mappingFile = null
         }