+++ CHANGELOG.md
@@ -1,3 +1,63 @@
+0.6.1 (released 14 September 2016)
+---------
+* Add runOnEachAssemble option to prevent task from running after assemble (#133)
+
+0.6.0 (released 31 August 2016)
+---------
+* Remove output annotations to opt out of Gradle caching (#132)
+* Add configurable slug for TeamCity statistics (#129)
+
+0.5.6 (released 17 August 2016)
+---------
+* Remove StyledTextOutput use to un-break Gradle 3.0 (#126)
+* Remove use of JDK-8-only methods on Process (#122)
+
+0.5.5 (released 30 June 2016)
+---------
+* Warn when building with Java 7 and below (#120)
+* Don't count Instant Run builds by default (#119)
+
+0.5.4 (released 15 June 2016)
+---------
+* Fix crash on old versions Gradle (#112)
+* Fix plugin for Gradle 2.14 (#110)
+
+0.5.3 (released 7 June 2016)
+---------
+* Fix crash when logging a crash (thanks, Gradle) (#105)
+* Add extra diagonistics when running with `--stacktrace` (#102)
+
+0.5.2 (released 17 May 2016)
+---------
+* Un-break for JVM 7 and below (PR #99)
+
+0.5.1 (released 16 May 2016)
+---------
+
+* Fix duplicate task invocation when `assemble` and `countDexMethods` are named on the command line (#97)
+* Fix crash when `chartDir` already exists (#94)
+* Fix crash when dx invocation fails for .aar (#93)
+* Fix JVM verifier crash, again (#86)
+
+0.5.0 (released 18 April 2016)
+---------
+
+* Add Android Test Module support (#81)
+* Add Instant Run support (#78)
+* Add AAR support (#70)
+* Add TeamCity support (#68)
+
+0.4.4 (released 4 April 2016)
+---------
+
+* Fix display of methods-remaining for multidex builds (don't show negative numbers) (issue #64)
+* Add 'maxTreeDepth' config option (issue #54)
+
+0.4.3 (released 14 March 2016)
+---------
+
+* Add sunburst chart generation (issue #57)
+
 0.4.2 (released 16 February 2016)
 ---------